Refactor (mostly mechanical): Extract class from LuisClient into LuisPredeictionClient
